requirements.txt
Django==2.0.1
djangorestframework==3.7.7
gitdb2==2.0.3
GitPython==2.1.8
pbr==3.1.1
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
smmap2==2.0.3
stevedore==1.28.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
